backend/server_beta/requirements.txt
django==2.0.5
psycopg2==2.7.4
django-oauth-toolkit==1.0.0
django-cors-middleware==1.3.1
djangorestframework==3.7.7
python-decouple==3.1
dj-database-url==0.5.0
django-filter==1.1.0
rest_condition==1.0.3
django-rest-swagger==2.2.0
django-storages==1.6.6
boto==2.48.0
mongoengine==0.15.0
pymongo==3.7.0
django-rest-framework-mongoengine-3.3.1

